% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmFE.R
\name{construct_weights}
\alias{construct_weights}
\title{Construct weigths}
\usage{
construct_weights(treatment, ps)
}
\arguments{
\item{treatment}{A matrix of treatments where rows correspond to units.}

\item{ps}{A matrix of propensity scores.}
}
\description{
Construct weigths
}
