% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmFE.R
\name{estimate_pscore}
\alias{estimate_pscore}
\title{Function to estimate propensity scores}
\usage{
estimate_pscore(formula, data, id_time_vec = NULL,
  bias_correct = FALSE, is_FE = TRUE, impute_FE = FALSE)
}
\description{
Function to estimate propensity scores
}
