% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmFE.R
\name{msmFE}
\alias{msmFE}
\title{Main function  for estimating treatment effects}
\usage{
msmFE(formula_outcome = NULL, formula_treatment, data, index,
  estimator = "HT", fixed_effects = TRUE, options = list())
}
\description{
Main function  for estimating treatment effects
}
